test: pair contract tests with the files they cover
The audit work added build_contract_test.go, lifecycle_contract_test.go,
and registration_contract_test.go, none of which had a corresponding
source file. The root package otherwise pairs every X_test.go with an
X.go, so these broke the convention.

Redistribute each test to the file implementing the behavior it covers:

- Build cancellation/context and all registration tests -> collection_test.go
  (Build and the registration APIs are on *collection)
- Disposable close dedup, concurrent close, close error aggregation, and
  root scope initializers -> provider_test.go (disposal tracking lives in
  provider.trackDisposable)
- Scope cancellation cleanup -> scope_test.go

No test bodies changed and no tests were dropped; the package still
exports the same 71 test functions. Also document the pairing rule in
CONTRIBUTING.md so it does not regress.
